refactor: [SRTP-815] move update valid tests in test_update
